refactor(frontend): move welcomeView and fileDecorations under src/frontend/ui
Both modules import `vscode` and perform VS Code UI wiring (registering a
TreeDataProvider for the welcome view and a FileDecorationProvider for
modified-file badges). Per the VS Code-zone guidance in CLAUDE.md,
platform-specific UI wiring belongs under `src/frontend/`.

Updated `src/extension.ts` to import them via the `@frontend/ui/*` alias.

Closes #3053
